Initial commit. Bare RTM for a no-HID display.
Very simple, very messy, and almost what I had in mind. I'd like it to fill the whole screen, but after trying everything you see in that bottom comment, I gave up and had the browser zoom in to the page. It is an inelegant and slap-dash script, but it works perfectly for my needs and doesn't break any of Remember The Milk's syncing or auto-update stuff, so I'm happy with it.